requirements.txt
# Azure SDK Dependencies
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Configuration and Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.0

# Logging and Monitoring
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
